Refactor TinyLlama and MicroLlama data class
